requirements.txt
MySQL-python
django==1.5.2
bae_image
bae_memcache
